# openapi.yml
openapi: 3.0.0
info:
  title: Sample API
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.1.9

servers:
  - url: http://api.example.com/v1
    description: Optional server description, e.g. Internal staging server for testing

paths:
  /kaimemo:
    get:
      summary: 買い物一覧取得
      description: 全件取得する
      responses:
        200: 
          $ref: '#/components/responses/GetKaimemos'
    post:
      summary: 買い物追加
      description: 買い物を追加する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                tag:
                  type: string
      responses:
        200:
          description: OK
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/GeneralError'
  /kaimemo/{id}:
    delete:
      summary: 買い物削除
      description: 買い物を削除する
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/GeneralError'
components:
  responses:
    GetKaimemos:
      description: 買い物一覧取得
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Kaimemo'
    UnauthorizedError:
      description: Access token is missing or invalid
    NotFoundError:
      description: The specified resource was not found
    GeneralError:
      description: Unexpected error
  schemas:
    Kaimemo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tag:
          type: string
        done:
          type: boolean